Add: implement http message and exceptions

+++ src/main/java/com/bazar/bazar/service/Impl/ProductServiceImpl.java
@@ -1,5 +1,7 @@
 package com.bazar.bazar.service.Impl;
 
+import com.bazar.bazar.exception.BadRequestException;
+import com.bazar.bazar.exception.NotFoundException;
 import com.bazar.bazar.model.Productos;
 import com.bazar.bazar.repository.ProductoRepository;
 import com.bazar.bazar.service.ProductService;
@@ -7,8 +9,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import javax.persistence.EntityNotFoundException;
+import java.util.Collections;
 import java.util.List;
-import java.util.Optional;
+
 
 @Service
 public class ProductServiceImpl implements ProductService {
@@ -17,32 +21,57 @@ public class ProductServiceImpl implements ProductService {
     private ProductoRepository productoRepository;
 
     @Override
-    public void crearProduct(Productos prod) {
-
-        Productos nuevoProducto = productoRepository.save(prod);
+    public Productos crearProduct(Productos prod) {
+        if (prod.getNomproduct().isEmpty()|| prod.getNomproduct()==null) {
+            throw new BadRequestException("nameProduct is invalid");
+        } else if ((prod.getPrecio() < 0)) {
+            throw new BadRequestException("precie is invalid");
+        } else if ((prod.getStock() < 0)) {
+            throw new BadRequestException("stock is invalid");
+        } else {
+            return productoRepository.save(prod);
+        }
     }
+
 
     @Override
     public List<Productos> listarProduct() {
 
         List<Productos> productos = productoRepository.findAll();
-        return productos;
+        if(productos.isEmpty()){
+            throw new NotFoundException("empty list");
+        }
+        else {
+            return productos;
+        }
     }
 
     @Override
     public Productos editarProductId(Productos productos, Long id) {
-        Productos proEditado = productoRepository.findById(id).orElseThrow();
-        proEditado.setNomproduct(productos.getNomproduct());
-        proEditado.setPrecio(productos.getPrecio());
-        proEditado.setStock(productos.getStock());
-        productoRepository.save(productos);
-        return productos;
+
+        Productos proEditado = productoRepository.findById(id).orElse(null);
+        if(proEditado==null){
+            throw new EntityNotFoundException( id + " not found in database");
+        }
+        else {
+            proEditado.setNomproduct(productos.getNomproduct());
+            proEditado.setPrecio(productos.getPrecio());
+            proEditado.setStock(productos.getStock());
+            productoRepository.save(productos);
+            return productos;
+        }
     }
 
     @Override
-    public void deletedProduct(Long id) {
-
-        productoRepository.deleteById(id);
+    public ResponseEntity deletedProduct(Long id) {
+        Productos product = productoRepository.findById(id).orElse(null);
+        if(product==null){
+            throw new EntityNotFoundException(id + " not found in database");
+        }
+        else {
+            productoRepository.deleteById(id);
+        }
+        return null;
     }
 
     @Override
@@ -56,6 +85,13 @@ public class ProductServiceImpl implements ProductService {
     public Productos productMasBarato() {
         List<Productos> productos = productoRepository.findAll();
         return productos.stream().min((a,b)-> a.getPrecio()- b.getPrecio()).get();
+    }
+
+    @Override
+    public List<Productos> ordenMenAMay() {
+        List<Productos> productos = productoRepository.findAll();
+         Collections.sort(productos);
+          return productos;
     }
